Move display component into components folder and container into container. Added SASS css to the login page.
