+++ frontend/src/Components/Homepage/Styles/Homepage.css
@@ -1,0 +1,38 @@
+.home {
+  display: grid;
+  grid-template-columns: auto 800px auto;
+  grid-template-rows: 70px auto;
+  grid-template-areas: ". login_menu ." ". signup_body ."; }
+  .home .login_menu {
+    grid-area: login_menu;
+    display: grid;
+    grid-template-columns: 240px auto 360px;
+    grid-template-areas: "logo . login_form"; }
+    .home .login_menu .logo {
+      grid-area: logo;
+      display: grid;
+      grid-template-columns: 70px 140px; }
+      .home .login_menu .logo img {
+        width: 70px;
+        height: 70px; }
+      .home .login_menu .logo h1 {
+        font-size: 2em;
+        align-self: center;
+        margin-top: 15px; }
+    .home .login_menu .login_form {
+      grid-area: login_form;
+      align-self: center; }
+  .home .signup_body {
+    grid-area: signup_body;
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: 1fr;
+    grid-template-areas: "about_us signup_form"; }
+    .home .signup_body .about_us {
+      grid-area: about_us; }
+    .home .signup_body .signup_form {
+      grid-area: signup_form;
+      align-self: center;
+      line-height: 3; }
+
+/*# sourceMappingURL=Homepage.css.map */
