+++ frontend/src/Components/RoomEaseContainer.jsx
@@ -2,14 +2,14 @@ import React, { Component } from "react"
 import { connect } from "react-redux" // ALLOWS THE USE OF REDUC TO GET THE STATE FROM THIS.PROPS
 import { Route, Switch } from "react-router-dom";
 
-import Homepage from "./Homepage/Homepage";
+import HomepageContainer from "./Homepage/Containers/HomepageContainer";
 import Setup from "./Setup/Setup";
 
 class RoomEaseContainer extends Component {
   render() {
     return(
       <Switch>
-        <Route exact path="/" component={Homepage} />
+        <Route exact path="/" component={HomepageContainer} />
         <Route path="/setup" component={Setup} />
       </Switch>
     )
